Fix the import error about fcrypt on WIndows
Fixed the import error when using Windows because of crypt module not existing there. Regarding the issue recently of the 'no file found dictionary.txt' or 'no file found dictionary.txt' You need to create both the file and put it into the same directory because the project's zip file does not include them. Any questions or problem, please comment.

Closes #8