Updated Brick-Out Game
Made some style changes (ball size, speed).  Sharpened the collision method to eliminate a brick when the ball hits the bottom of the brick exactly, instead of at its center point.  Still need to fine-tune the collision point at the side and top of the brick.  Added a random x-axis placement for the beginning of the game.  Added some comments to help with debugging.

I'm new to GitHub, so please be direct with responses.