game of craps.py
this is a dice game where when we follow a certain set of instructions and the sum of numbers of rolled dices determine whether you have won or not. You roll two six-sided dice, each with faces containing one, two, three, four, five and six spots, respectively. When the dice come to rest, the sum of the spots on the two front faces is calculated. If the sum is 7 or 11 on the first roll, you win. If the sum is 2, 3 or 12 on first roll, you lose. If the sum is 4, 5, 6, 8, 9 or 10 on the first roll, that sum becomes your “point”. To win, you must continue rolling the dice until you “make your point”. You lose by rolling a 7 before making your point