Update youtubedownloader.py
The use of global variables should be minimized as much as possible. In the current code, url_box is a global variable. Instead, you can pass the url_box as a parameter to the download function.
It is good practice to include exception handling in your code, especially when working with web requests. You can add a try-except block around the line url = YouTube(str(url_box.get())) to handle any errors that may occur.
You can add a progress bar to show the progress of the download. This will give the user a better idea of how much time is left until the download is complete.
You can also add a feature to let the user choose the download location instead of downloading to the default directory.