Update FizzBuzz.py
logical issue in the FizzBuzz() function.

The current implementation of the FizzBuzz() function requires the user to input the number to count up to. This limits the functionality of the function to a single use case.

To make the function more flexible and reusable, we can modify the function to take the num argument as an input parameter.                                                                                 With this modification, we can call the FizzBuzz() function with any number you want to count up to.